Unify get_block_size (#3039)
* Unify get_block_size

* Remove granularity defines in the pt2e path

* Fix format
