+++ torchao/quantization/__init__.py
@@ -19,6 +19,7 @@ from .GPTQ import (
     MultiTensorInputRecorder,
 )
 from .granularity import (
+    Granularity,
     PerAxis,
     PerGroup,
     PerRow,
@@ -197,6 +198,7 @@ __all__ = [
     "MappingType",
     "ZeroPointDomain",
     "TorchAODType",
+    "Granularity",
     "PerTensor",
     "PerAxis",
     "PerGroup",
